Fixed check features
