+++ module-system/module-implementations/sov-chain-state/src/lib.rs
@@ -145,6 +145,29 @@ impl<C: sov_modules_api::Context, Cond: ValidityCondition> ChainState<C, Cond> {
             .get(working_set)
             .expect("Slot height should be set at initialization")
     }
+
+    /// Return the genesis hash of the module.
+    pub fn get_genesis_hash(&self, working_set: &mut WorkingSet<C::Storage>) -> Option<[u8; 32]> {
+        self.genesis_hash.get(working_set)
+    }
+
+    /// Returns the transition in progress of the module.
+    pub fn get_in_progress_transition(
+        &self,
+        working_set: &mut WorkingSet<C::Storage>,
+    ) -> Option<TransitionInProgress<Cond>> {
+        self.in_progress_transition.get(working_set)
+    }
+
+    /// Returns the completed transition associated with the provided `transition_num`.
+    pub fn get_historical_transitions(
+        &self,
+        transition_num: TransitionHeight,
+        working_set: &mut WorkingSet<C::Storage>,
+    ) -> Option<StateTransitionId<Cond>> {
+        self.historical_transitions
+            .get(&transition_num, working_set)
+    }
 }
 
 impl<C: sov_modules_api::Context, Cond: ValidityCondition> sov_modules_api::Module
