+++ module-system/module-implementations/sov-chain-state/src/query.rs
@@ -3,8 +3,7 @@ use sov_modules_api::macros::rpc_gen;
 use sov_rollup_interface::zk::ValidityCondition;
 use sov_state::WorkingSet;
 
-use super::ChainState;
-use crate::{StateTransitionId, TransitionHeight, TransitionInProgress};
+use crate::{ChainState, TransitionHeight};
 
 #[rpc_gen(client, server, namespace = "chainState")]
 impl<C: sov_modules_api::Context, Cond: ValidityCondition> ChainState<C, Cond> {
@@ -16,28 +15,5 @@ impl<C: sov_modules_api::Context, Cond: ValidityCondition> ChainState<C, Cond> {
         working_set: &mut WorkingSet<C::Storage>,
     ) -> RpcResult<TransitionHeight> {
         Ok(self.get_slot_height(working_set))
-    }
-
-    /// Return the genesis hash of the module.
-    pub fn get_genesis_hash(&self, working_set: &mut WorkingSet<C::Storage>) -> Option<[u8; 32]> {
-        self.genesis_hash.get(working_set)
-    }
-
-    /// Returns the transition in progress of the module.
-    pub fn get_in_progress_transition(
-        &self,
-        working_set: &mut WorkingSet<C::Storage>,
-    ) -> Option<TransitionInProgress<Cond>> {
-        self.in_progress_transition.get(working_set)
-    }
-
-    /// Returns the completed transition associated with the provided `transition_num`.
-    pub fn get_historical_transitions(
-        &self,
-        transition_num: TransitionHeight,
-        working_set: &mut WorkingSet<C::Storage>,
-    ) -> Option<StateTransitionId<Cond>> {
-        self.historical_transitions
-            .get(&transition_num, working_set)
     }
 }
